# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organisations/{organisationId}": {
      "get": {
        "operationId": "Organisations_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Organisation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organisation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Organisations"
        ],
        "x-required-permissions": [
          "org:read"
        ],
        "summary": "Get the organisation",
        "description": "**Requires permission:** `org:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/me": {
      "get": {
        "operationId": "Organisations_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgMe"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Organisations"
        ],
        "x-required-permissions": [
          "org:read"
        ],
        "summary": "The caller's organisation role, reach, and explicit practice/company grants",
        "description": "**Requires permission:** `org:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/roles": {
      "get": {
        "operationId": "Organisations_roles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "role:read"
        ],
        "summary": "List system and custom roles + the permission catalogue",
        "description": "**Requires permission:** `role:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      },
      "post": {
        "operationId": "Organisations_createRole",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Makes retries safe: a repeated request with the same key and body replays the first response (24h).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "role:manage"
        ],
        "summary": "Create a custom role (cannot exceed the creator's own permissions)",
        "description": "**Requires permission:** `role:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{organisationId}/members": {
      "get": {
        "operationId": "Organisations_members",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Members",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "member:read"
        ],
        "summary": "List members",
        "description": "**Requires permission:** `member:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/members/{membershipId}": {
      "patch": {
        "operationId": "Organisations_updateMember",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "member:manage"
        ],
        "summary": "Change a member's role, company scope or status",
        "description": "**Requires permission:** `member:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRequest"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "Organisations_removeMember",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "member:manage"
        ],
        "summary": "Remove a member (the last owner cannot be removed)",
        "description": "**Requires permission:** `member:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/invitations": {
      "post": {
        "operationId": "Organisations_invite",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Makes retries safe: a repeated request with the same key and body replays the first response (24h).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Invitation sent asynchronously",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invitation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "member:invite"
        ],
        "summary": "Invite a person by email",
        "description": "**Requires permission:** `member:invite`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "Organisations_invitations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Invitations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "member:read"
        ],
        "summary": "List pending invitations",
        "description": "**Requires permission:** `member:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/invitations/{invitationId}": {
      "delete": {
        "operationId": "Organisations_revoke",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Access control"
        ],
        "x-required-permissions": [
          "member:invite"
        ],
        "summary": "Revoke an invitation",
        "description": "**Requires permission:** `member:invite`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/invitations/accept": {
      "post": {
        "operationId": "Organisations_accept",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Joined"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Access control"
        ],
        "summary": "Accept an invitation (caller must be the invited, verified address)",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "UK Accounting Platform API",
    "description": "Practice-first accounting, tax and compliance platform. **V0 foundation:** identity, tenancy, RBAC, documents, jobs, audit, outbox events, workflows, tasks, notifications, integration and AI abstractions.\n\nAll tenant data lives under `/organisations/{organisationId}` and is protected by authentication, RBAC, company scope and PostgreSQL row-level security. Errors use RFC 9457 `application/problem+json`.",
    "version": "1.0.0",
    "contact": {},
    "x-api-versioning": "URI versioning (/api/v1). Additive changes only within a version; breaking changes ship as /api/v2."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Opaque session token from `POST /auth/login/bearer`"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uk_session",
        "description": "httpOnly SameSite=Strict cookie set by `POST /auth/login`. Unsafe methods must send an allowed `Origin`."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "urn:uk-platform:error:validation_error"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "instance": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "description": "Echoes `x-request-id`"
          }
        }
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isSystem": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "description",
          "permissions",
          "isSystem"
        ],
        "additionalProperties": false
      },
      "Invitation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "expiresAt"
        ],
        "additionalProperties": false
      },
      "Organisation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRACTICE",
              "BUSINESS"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "status"
        ],
        "additionalProperties": false
      },
      "OrgMe": {
        "type": "object",
        "properties": {
          "organisationId": {
            "type": "string",
            "format": "uuid"
          },
          "organisationType": {
            "type": "string",
            "enum": [
              "PRACTICE",
              "BUSINESS"
            ]
          },
          "role": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Organisation-level role permissions; company/practice decisions use grants"
          },
          "companyScope": {
            "type": "string",
            "enum": [
              "ALL",
              "ASSIGNED"
            ],
            "description": "Reach of the organisation role: ALL = organisation-wide default"
          },
          "companyIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Companies with an explicit company-level grant"
          },
          "practiceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Practices with an explicit practice-level grant"
          }
        },
        "required": [
          "organisationId",
          "organisationType",
          "role",
          "permissions",
          "companyScope",
          "companyIds",
          "practiceIds"
        ],
        "additionalProperties": false
      },
...
      "RoleList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "key": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "permissions": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "isSystem": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "key",
                "name",
                "description",
                "permissions",
                "isSystem"
              ],
              "additionalProperties": false
            }
          },
          "permissionCatalogue": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "items",
          "permissionCatalogue"
        ],
        "additionalProperties": false
      },
      "MemberList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "displayName"
                  ],
                  "additionalProperties": false
                },
                "role": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "key": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "key",
                    "name"
                  ],
                  "additionalProperties": false
                },
                "status": {
                  "type": "string"
                },
                "companyScope": {
                  "type": "string",
                  "enum": [
                    "ALL",
                    "ASSIGNED"
                  ]
                },
                "companyIds": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "companyGrants": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "companyId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "role": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "key": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "key"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "companyId",
                      "role"
                    ],
                    "additionalProperties": false
                  }
                },
                "practiceGrants": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "practiceId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "role": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "key": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "key"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "practiceId",
                      "role"
                    ],
                    "additionalProperties": false
                  }
                }
              },
              "required": [
                "id",
                "user",
                "role",
                "status",
                "companyScope",
                "companyIds",
                "companyGrants",
                "practiceGrants"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
      "InvitationList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "email": {
                  "type": "string"
                },
                "expiresAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "email",
                "expiresAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
      "CreateRoleRequest": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{2,40}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "description": {
            "type": "string",
            "maxLength": 300,
            "default": ""
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "org:read",
                "org:manage",
                "member:read",
                "member:invite",
                "member:manage",
                "role:read",
                "role:manage",
                "company:read",
                "company:create",
                "company:update",
                "period:read",
                "period:manage",
                "document:read",
                "document:upload",
                "document:archive",
                "audit:read",
                "job:read",
                "job:manage",
                "task:read",
                "task:manage",
                "workflow:read",
                "workflow:manage",
                "integration:read",
                "integration:manage",
                "ai:use",
                "ai:approve",
                "practice:read",
                "practice:manage",
                "practice:member:manage",
                "company:access:manage",
                "workflow:review",
                "workflow:approve"
              ]
            },
            "minItems": 1
          }
        },
        "required": [
          "key",
          "name",
          "permissions"
        ],
        "additionalProperties": false
      },
      "UpdateMemberRequest": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "companyScope": {
            "type": "string",
            "enum": [
              "ALL",
              "ASSIGNED"
            ]
          },
          "companyIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 500
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED"
            ]
          }
        },
        "additionalProperties": false
      },
      "InviteMemberRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "companyScope": {
            "type": "string",
            "enum": [
              "ALL",
              "ASSIGNED"
            ],
            "default": "ALL"
          },
          "companyIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 500,
            "default": []
          }
        },
        "required": [
          "email",
          "roleId"
        ],
        "additionalProperties": false
      },
      "AcceptInvitationRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 10,
            "maxLength": 200
          }
        },
        "required": [
          "token"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "400": {
        "description": "Malformed request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "401": {
        "description": "Not authenticated or session expired",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "403": {
        "description": "Authenticated but not permitted (RBAC, segregation of duties, CSRF origin)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "404": {
        "description": "Not found — also returned for resources in other organisations (existence is never revealed)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "422": {
        "description": "Validation failed (`errors[]` lists fields)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}